+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Renamed type of recipes to mode to avoid confusion with item types.
+- Only use one sub-directory for icons instead of two. 
 
 ## 1.0.0 - 2018-02-19